Refactor Checklist folder into subfolders
